improve project structure + update readme
